JOJ UFO + snarks + misc. fixes
- Added JOJ UFO.
- Added Snarks.
- Added Snark Nests.
- Fixed mouth movement for Chavs (it's intended to move sideways, like a broken jaw)
- Fixed small issues with [REDACTED].

Bad news:
There is a small number of unused/cut content in Crack-Life. Guardchav is one of them: https://www.youtube.com/watch?v=MUDMSzMjAdg
Unfortunately files for it are not present in Crack-Life: Campaign Edition (except for a leftover animation of a Chav transforming in to Guardchav, as seen in the video). According to siemka321, Guardchav is lost, since the original files are no longer in his computer. If anyone in this world somehow reads this and still has some leftover files for it (qc, smds (or full compiled mdl) + sounds), please let me know.